Changes: - Created playback_routes.py, all endpoints used by playback.html that was originally in app.py is now in a separate document.
